src: move registers into their own module
